Fix debug_draw library integration
