--- Engine/UI/IDebugDrawOpenGLImplementation.cpp
@@ -1,11 +1,0 @@
-#include "IDebugDrawOpenGLImplementation.h"
-
-IDebugDrawOpenGLImplementation::IDebugDrawOpenGLImplementation()
-{
-
-}
-
-IDebugDrawOpenGLImplementation::~IDebugDrawOpenGLImplementation()
-{
-
-}
