--- Engine/UI/IDebugDrawOpenGLImplementation.h
@@ -1,14 +1,0 @@
-#ifndef _IDEBUGDRAWOPENGLIMPLEMENTATION_H_
-#define _IDEBUGDRAWOPENGLIMPLEMENTATION_H_
-
-
-class IDebugDrawOpenGLImplementation
-{
-public:
-	IDebugDrawOpenGLImplementation();
-	~IDebugDrawOpenGLImplementation();
-
-};
-
-#endif //_IDEBUGDRAWOPENGLIMPLEMENTATION_H_
-
